--- cg/meta/orders/microsalt_submitter.py
@@ -1,5 +1,0 @@
-from cg.meta.orders.microbial_submitter import MicrobialSubmitter
-
-
-class MicrosaltSubmitter(MicrobialSubmitter):
-    pass
