+++ cg/services/orders/store_order_services/store_fastq_order_service.py
@@ -23,9 +23,7 @@ class StoreFastqOrderService(StoreOrderService):
     def store_order(self, order: OrderIn) -> dict:
         """Submit a batch of samples for FASTQ delivery."""
 
-        project_data, lims_map = self.lims(
-            lims_api=self.lims, lims_order=order, new_samples=order.samples
-        )
+        project_data, lims_map = self.lims.process_lims(lims_order=order, new_samples=order.samples)
         status_data = self.order_to_status(order)
         self._fill_in_sample_ids(samples=status_data["samples"], lims_map=lims_map)
         new_samples = self.store_items_in_status(
@@ -158,4 +156,4 @@ class StoreFastqOrderService(StoreOrderService):
         for sample_obj in samples:
             LOG.info(f"{sample_obj.internal_id}: add missing reads in LIMS")
             target_reads = sample_obj.application_version.application.target_reads / 1000000
-            self.lims.update_sample(sample_obj.internal_id, target_reads=target_reads)
+            self.lims.lims_api.update_sample(sample_obj.internal_id, target_reads=target_reads)
